Сделал компоненту controller/board
