+++ src/main.js
@@ -1,12 +1,7 @@
 import SiteMenuComponent from "./components/site-menu";
-import SortComponent from "./components/sort";
 import FilterComponent from "./components/site-filter";
-import TaskComponent from "./components/task-elem";
-import TasksComponent from "./components/tasks.js";
-import TaskEditComponent from "./components/task-edit-elem";
-import NoTasksComponent from "./components/no-tasks.js";
-import LoadMoreButtonComponent from "./components/button-load-more";
 import BoardComponent from "./components/board";
+import BoardController from "./controllers/board.js";
 import {
   generateTasks
 } from "./mock/task";
@@ -14,9 +9,7 @@ import {
   generateFilters
 } from "./mock/site-filter";
 import {
-  render,
-  replace,
-  remove
+  render
 } from './utils/render';
 import {
   renderPosition
@@ -24,78 +17,6 @@ import {
 
 
 const TASK_COUNT = 22;
-const SHOWING_TASKS_COUNT_ON_START = 8;
-const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
-
-const renderTask = (taskListElement, task) => {
-  const replaceTaskToEdit = () => {
-    replace(taskEditComponent, taskComponent);
-  };
-
-  const replaceEditToTask = () => {
-    replace(taskComponent, taskEditComponent);
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      replaceEditToTask();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const taskComponent = new TaskComponent(task);
-  const taskEditComponent = new TaskEditComponent(task);
-  taskComponent.setEditButtonClickHandler(() => {
-    replaceTaskToEdit();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  taskEditComponent.setSubmitHandler((evt) => {
-    evt.preventDefault();
-    replaceEditToTask();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
-
-  render(taskListElement, taskComponent, renderPosition.BEFOREEND);
-};
-
-const renderBoard = (boardComponent, tasks) => {
-  const isAllTasksArchived = tasks.every((task) => task.isArchive);
-
-  if (isAllTasksArchived) {
-    render(boardComponent.getElement(), new NoTasksComponent(), renderPosition.BEFOREEND);
-    return;
-  }
-
-  render(boardComponent.getElement(), new SortComponent(), renderPosition.BEFOREEND);
-  render(boardComponent.getElement(), new TasksComponent(), renderPosition.BEFOREEND);
-
-  const taskListElement = boardComponent.getElement().querySelector(`.board__tasks`);
-
-  let showingTasksCount = SHOWING_TASKS_COUNT_ON_START;
-  tasks.slice(0, showingTasksCount)
-    .forEach((task) => {
-      renderTask(taskListElement, task);
-    });
-
-  const loadMoreButtonComponent = new LoadMoreButtonComponent();
-
-  render(boardComponent.getElement(), loadMoreButtonComponent, renderPosition.BEFOREEND);
-
-  loadMoreButtonComponent.setClickHandler(() => {
-    const prevTasksCount = showingTasksCount;
-    showingTasksCount = showingTasksCount + SHOWING_TASKS_COUNT_BY_BUTTON;
-
-    tasks.slice(prevTasksCount, showingTasksCount)
-      .forEach((task) => renderTask(taskListElement, task));
-
-    if (showingTasksCount >= tasks.length) {
-      remove(loadMoreButtonComponent);
-    }
-  });
-};
 
 const siteMainElement = document.querySelector(`.main`);
 const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
@@ -107,6 +28,7 @@ render(siteHeaderElement, new SiteMenuComponent(), renderPosition.BEFOREEND);
 render(siteMainElement, new FilterComponent(filters), renderPosition.BEFOREEND);
 
 const boardComponent = new BoardComponent();
-render(siteMainElement, boardComponent, renderPosition.BEFOREEND);
+const boardController = new BoardController(boardComponent);
 
-renderBoard(boardComponent, tasks);
+render(siteMainElement, boardComponent, renderPosition.BEFOREEND);
+boardController.render(tasks);
